backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pymongo==4.13.2
beautifulsoup4==4.12.3
requests==2.31.0
Pillow==10.2.0
pytesseract==0.3.10
pyyaml==6.0.1
bson==0.5.10
google-generativeai==0.3.2 